#59 #60  Moved ActionResourceInfo object to models
